# [Performance] Make determining whether a code point represents a combining mark faster