@zyedidia 2020-06-11
I don't think we should cut corners on the unicode mark range, but we can definitely look into ways to speed this up. One of the reasons to implement combining characters was so that rendering Thai script would be correct (#872), and this change would break that.

A good method of improving performance would be to check if the rune is a common character (ASCII) that is known not to be a mark before doing an expensive check to see if it is a mark. I've looked closer at the Go documentation and see there is a `unicode.IsMark` function, and inspecting the source code seems to reveal that this is optimized to handle the common case of Latin characters quickly (I think), by making sure that the rune is not Latin before doing anything expensive. Let's try using that function and see if that improves performance.