@p-e-w 2020-06-12
Thanks, I somehow managed to completely overlook `unicode.IsMark`. That function indeed has a fast path, but we can do better. `unicode.IsMark` only skips testing against the full range table if the rune is less than 0x00FF, but the first mark only occurs at 0x0300.

So that is the fast path I ended up implementing. For Latin-1 characters, this code is actually even faster than the previous one from this PR, as it only needs to perform a single comparison instead of five.

Speed improvement is similar to before (note that `~` does not mean that there has been no improvement, but that the p-value returned by the statistical test is above the significance threshold):

```
name                           old time/op  new time/op  delta
CreateAndClose10Lines-4        41.3µs ± 1%  47.4µs ±33%     ~     (p=0.700 n=3+3)
CreateAndClose100Lines-4        122µs ± 4%   128µs ±13%     ~     (p=1.000 n=3+3)
CreateAndClose1000Lines-4       897µs ± 2%   877µs ± 2%     ~     (p=0.400 n=3+3)
CreateAndClose10000Lines-4     9.14ms ± 2%  8.91ms ± 0%   -2.44%  (p=0.100 n=3+3)
CreateAndClose100000Lines-4    86.4ms ± 5%  80.0ms ± 1%   -7.45%  (p=0.100 n=3+3)
CreateAndClose1000000Lines-4    825ms ± 7%   802ms ± 0%     ~     (p=0.700 n=3+3)
Read10Lines-4                  1.90µs ± 3%  1.91µs ± 1%     ~     (p=1.000 n=3+3)
Read100Lines-4                 12.5µs ± 5%  12.4µs ± 1%     ~     (p=1.000 n=3+3)
Read1000Lines-4                 122µs ± 6%   119µs ± 1%     ~     (p=1.000 n=3+3)
Read10000Lines-4               1.47ms ± 9%  1.39ms ± 1%     ~     (p=0.700 n=3+3)
Read100000Lines-4              14.3ms ± 4%  14.3ms ± 3%     ~     (p=1.000 n=3+3)
Read1000000Lines-4              148ms ± 8%   131ms ± 6%     ~     (p=0.200 n=3+3)
Edit10Lines1Cursor-4            245µs ± 3%   158µs ± 1%  -35.50%  (p=0.100 n=3+3)
Edit100Lines1Cursor-4           120µs ± 4%    80µs ± 1%  -33.63%  (p=0.100 n=3+3)
Edit100Lines10Cursors-4        6.50ms ± 3%  5.01ms ± 2%  -22.87%  (p=0.100 n=3+3)
Edit1000Lines1Cursor-4         65.5µs ± 3%  43.7µs ± 1%  -33.33%  (p=0.100 n=3+3)
Edit1000Lines10Cursors-4       5.45ms ± 2%  4.14ms ± 0%  -24.11%  (p=0.100 n=3+3)
Edit1000Lines100Cursors-4       352ms ± 3%   269ms ± 0%  -23.67%  (p=0.100 n=3+3)
Edit10000Lines1Cursor-4         248µs ± 5%   185µs ± 4%  -25.34%  (p=0.100 n=3+3)
Edit10000Lines10Cursors-4      5.49ms ± 1%  4.37ms ± 1%  -20.40%  (p=0.100 n=3+3)
Edit10000Lines100Cursors-4      393ms ± 3%   302ms ± 0%  -23.18%  (p=0.100 n=3+3)
Edit10000Lines1000Cursors-4     33.4s ± 3%   26.1s ± 4%  -22.09%  (p=0.100 n=3+3)
Edit100000Lines1Cursor-4       2.30ms ±24%  2.11ms ±12%     ~     (p=0.400 n=3+3)
Edit100000Lines10Cursors-4     43.7ms ±24%  40.2ms ±15%     ~     (p=0.400 n=3+3)
Edit100000Lines100Cursors-4     707ms ± 1%   647ms ± 2%   -8.50%  (p=0.100 n=3+3)
Edit100000Lines1000Cursors-4    35.3s ± 5%   29.7s ± 5%  -15.82%  (p=0.100 n=3+3)
Edit1000000Lines1Cursor-4       433µs ± 1%   396µs ± 2%   -8.41%  (p=0.100 n=3+3)
Edit1000000Lines10Cursors-4     366ms ± 6%   404ms ±26%     ~     (p=1.000 n=3+3)
Edit1000000Lines100Cursors-4    4.17s ± 3%   4.22s ±12%     ~     (p=0.700 n=3+3)
Edit1000000Lines1000Cursors-4   74.8s ± 1%   68.8s ± 2%   -8.05%  (p=0.100 n=3+3)
```